@@ -32,23 +30,26 @@ allprojects {
             url "https://jitpack.io"
         }
         flatDir {
-            dirs 'ForgeModule/libs'
+            dirs "ForgeModule/libs"
         }
     }
 }
 
 ext {
-    compileSdkVersion = 26
-    buildToolsVersion = "26.0.1"
-    supportLibraryVersion = "26.0.1"
-    crosswalkVersion = "23.53.589.4"
+    dependencies = [
+        "androidx.legacy:legacy-support-v4:1.0.0",
+        "com.google.guava:guava:28.0-android",
+    ]
+
+    buildToolsVersion = "29.0.0"
 
     minSdkVersion = moduleMinSdkVersion() ?: 14
-    targetSdkVersion = 26
+    compileSdkVersion = 29
+    targetSdkVersion = 29
 
     compileOptions = {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     debug = {
